loadgen: Clean up behavior and documentation of outstanding queries.
* Remove server scenario hueristic for "max_oustanding_queries",
  which would spuriously early out for very low QPS and would
  be uselessly high for very large QPS.
* Add a TestSettings::server_max_async_queries that defaults
  to allowing unlimitted oustanding queries.
* Update documentation for the Server scenario, to explicitly
  state that there is no limit to the number of oustanding queries
  allowed.
* Small updates to other comments.


Former-commit-id: e4cc4bb44897b94fd13f7af531d73a6f8bf4336e

+++ loadgen/test_settings_internal.h
@@ -56,8 +56,8 @@ struct TestSettingsInternal {
   int samples_per_query;
   double target_qps;
   std::chrono::nanoseconds target_latency{0};
-  double target_latency_percentile;  // Single, multistream and server mode
-  int max_async_queries;
+  double target_latency_percentile;  // Single, multistream, and server modes.
+  uint64_t max_async_queries;
 
   // Target duration is used to generate queries of a minimum duration before
   // the test run.
